Layout.CreateFromString with throwConfigExceptions-parameter. Moved FuncLayoutRenderer-tests to SimpleLayoutParserTests
