+++ tests/NLog.UnitTests/ConfigFileLocatorTests.cs
@@ -43,11 +43,9 @@ namespace NLog.UnitTests
     using Microsoft.CSharp;
     using Xunit;
     using NLog.Config;
-    using NLog.Layouts;
-    using NLog.LayoutRenderers;
     using NLog.UnitTests.Mocks;
 
-    public class ConfigFileLocatorTests : NLogTestBase, IDisposable
+    public sealed class ConfigFileLocatorTests : NLogTestBase, IDisposable
     {
         private readonly string _tempDirectory;
 
@@ -61,91 +59,6 @@ namespace NLog.UnitTests
         {
             if (Directory.Exists(_tempDirectory))
                 Directory.Delete(_tempDirectory, true);
-        }
-
-        [Fact]
-        void FuncLayoutRendererRegisterTest1()
-        {
-            LayoutRenderer.Register("the-answer", (info) => "42");
-            Layout l = "${the-answer}";
-            var result = l.Render(LogEventInfo.CreateNullEvent());
-            Assert.Equal("42", result);
-        }
-
-        [Fact]
-        void FuncLayoutRendererFluentMethod_ThreadSafe_Test()
-        {
-            // Arrange
-            var layout = Layout.CreateFromMethod(l => "42", LayoutRenderOptions.ThreadSafe);
-            // Act
-            var result = layout.Render(LogEventInfo.CreateNullEvent());
-            // Assert
-            Assert.Equal("42", result);
-            Assert.True(layout.ThreadSafe);
-            Assert.False(layout.ThreadAgnostic);
-        }
-
-        [Fact]
-        void FuncLayoutRendererFluentMethod_ThreadAgnostic_Test()
-        {
-            // Arrange
-            var layout = Layout.CreateFromMethod(l => "42", LayoutRenderOptions.ThreadAgnostic);
-            // Act
-            var result = layout.Render(LogEventInfo.CreateNullEvent());
-            // Assert
-            Assert.Equal("42", result);
-            Assert.True(layout.ThreadSafe);
-            Assert.True(layout.ThreadAgnostic);
-        }
-
-        [Fact]
-        void FuncLayoutRendererFluentMethod_ThreadUnsafe_Test()
-        {
-            // Arrange
-            var layout = Layout.CreateFromMethod(l => "42", LayoutRenderOptions.None);
-            // Act
-            var result = layout.Render(LogEventInfo.CreateNullEvent());
-            // Assert
-            Assert.Equal("42", result);
-            Assert.False(layout.ThreadSafe);
-            Assert.False(layout.ThreadAgnostic);
-        }
-
-        [Fact]
-        void FuncLayoutRendererFluentMethod_NullThrows_Test()
-        {
-            // Arrange
-            Assert.Throws<ArgumentNullException>(() => Layout.CreateFromMethod(null));
-        }
-
-        [Fact]
-        void FuncLayoutRendererRegisterTest1WithXML()
-        {
-            LayoutRenderer.Register("the-answer", (info) => 42);
-
-            LogManager.Configuration = XmlLoggingConfiguration.CreateFromXmlString(@"
-<nlog throwExceptions='true'>
-            
-                <targets>
-                    <target name='debug' type='Debug' layout= 'TheAnswer=${the-answer:Format=D3}' /></targets>
-                <rules>
-                    <logger name='*' minlevel='Debug' writeTo='debug' />
-                </rules>
-            </nlog>");
-
-            var logger = LogManager.GetCurrentClassLogger();
-            logger.Debug("test1");
-            AssertDebugLastMessage("debug", "TheAnswer=042");
-        }
-
-        [Fact]
-        void FuncLayoutRendererRegisterTest2()
-        {
-            LayoutRenderer.Register("message-length", (info) => info.Message.Length);
-            Layout l = "${message-length}";
-            var result = l.Render(LogEventInfo.Create(LogLevel.Error, "logger-adhoc", "1234567890"));
-            Assert.Equal("10", result);
-
         }
 
         [Fact]
